[TEST] Add SimpleTestDevice
- Add simple_dev make target
- Add SimpleTestDevice to runtime_util
- Move public functions in general_dev to virtual_device_util
- Use virtual_device_util functions in General and Simple test devices
